add longest substring >= k repeats

+++ src/LongestSubstringNoRepeat.h
@@ -31,7 +31,7 @@ public:
 
         }
 
-        return maxLen;
+        return (int)maxLen;
     }
 
     int lengthOfLongestSubstring2(string s) {
